+++ nano/node/bootstrap_ascending/database_scan.cpp
@@ -1,6 +1,6 @@
 #include <nano/lib/utility.hpp>
+#include <nano/node/bootstrap_ascending/crawlers.hpp>
 #include <nano/node/bootstrap_ascending/database_scan.hpp>
-#include <nano/node/bootstrap_ascending/iterators.hpp>
 #include <nano/secure/common.hpp>
 #include <nano/secure/ledger.hpp>
 #include <nano/secure/ledger_set_any.hpp>
